+++ bottle.py
@@ -529,18 +529,22 @@ class Route:
         """ Return the callback. If the callback is a decorated function, try to
             recover the original function. """
         func = self.callback
-        func = getattr(func, '__func__', func)
-        while hasattr(func, '__closure__') and getattr(func, '__closure__'):
-            attributes = getattr(func, '__closure__')
-            func = attributes[0].cell_contents
-
-            # in case of decorators with multiple arguments
-            if not isinstance(func, FunctionType):
-                # pick first FunctionType instance from multiple arguments
-                func = filter(lambda x: isinstance(x, FunctionType),
-                              map(lambda x: x.cell_contents, attributes))
-                func = list(func)[0]  # py3 support
-        return func
+        while True:
+            if getattr(func, '__wrapped__', False):
+                func = func.__wrapped__
+            elif getattr(func, '__func__', False):
+                func = func.__func__
+            elif getattr(func, '__closure__', False):
+                depr(0, 14, "Decorated callback without __wrapped__",
+                     "When applying decorators to route callbacks, make sure" \
+                     " the decorator uses @functools.wraps or update_wrapper." \
+                     " This warning may also trigger if you reference callables" \
+                     " from a nonlocal scope.")
+                cells_values = (cell.cell_contents for cell in func.__closure__)
+                isfunc = lambda x: isinstance(x, FunctionType) or hasattr(x, '__call__')
+                func = next(filter(isfunc, cells_values), func)
+            else:
+                return func
 
     def get_callback_args(self):
         """ Return a list of argument names the callback (most likely) accepts
@@ -561,7 +565,9 @@ class Route:
 
     def __repr__(self):
         cb = self.get_undecorated_callback()
-        return '<%s %s -> %s:%s>' % (self.method, self.rule, cb.__module__, cb.__name__)
+        return '<%s %s -> %s:%s>' % (
+            self.method, self.rule, cb.__module__, getattr(cb, '__name__', '?')
+        )
 
 ###############################################################################
 # Application Object ###########################################################